cr + pr: #DRS-163 Now listening to transaction commit events

+++ drs/src/core/com/db4o/drs/foundation/Signatures.java
@@ -27,7 +27,7 @@ public class Signatures {
 		_bySignature.put(signature, databaseId);
 	}
 
-	public long idFor(DrsUUID uuid) {
+	public int idFor(DrsUUID uuid) {
 		return idFor(new Signature(uuid.getSignaturePart()));
 	}
